Fix for DIRSTUDIO-372: Added null checks
git-svn-id: https://svn.apache.org/repos/asf/directory/studio/trunk@687753 13f79535-47bb-0310-9956-ffa450edef68

+++ ldapbrowser-ui/src/main/java/org/apache/directory/studio/ldapbrowser/ui/views/modificationlogs/ModificationLogsViewUniversalListener.java
@@ -141,14 +141,14 @@ public class ModificationLogsViewUniversalListener implements EntryUpdateListene
     void setInput( ModificationLogsViewInput input )
     {
         // only if another connection is selected
-        if ( this.input != input )
+        if ( this.input != input && input.getBrowserConnection().getConnection() != null )
         {
             this.input = input;
 
             // load file %u %g
             StringBuffer sb = new StringBuffer();
             LdifModificationLogger modificationLogger = ConnectionCorePlugin.getDefault().getLdifModificationLogger();
-            File[] files = modificationLogger.getFiles( input.getConnection().getConnection() );
+            File[] files = modificationLogger.getFiles( input.getBrowserConnection().getConnection() );
             int i = input.getIndex();
             if ( 0 <= i && i < files.length && files[i] != null && files[i].exists() && files[i].canRead() )
             {
@@ -225,34 +225,36 @@ public class ModificationLogsViewUniversalListener implements EntryUpdateListene
 
 
     /**
-     * Clears the input and deletes the logfiles for it
-     * 
+     * Clears the input and deletes the logfiles for it.
      */
     public void clearInput()
     {
-        StringBuffer sb = new StringBuffer( "" );
-        FileWriter fw = null;
-        LdifModificationLogger modificationLogger = ConnectionCorePlugin.getDefault().getLdifModificationLogger();
-        File[] files = modificationLogger.getFiles( input.getConnection().getConnection() );
-        modificationLogger.dispose( input.getConnection().getConnection() );
-        for ( int i = 0; i < files.length; i++ )
+        if ( input.getBrowserConnection().getConnection() != null )
         {
-            try
+            StringBuffer sb = new StringBuffer( "" );
+            FileWriter fw = null;
+            LdifModificationLogger modificationLogger = ConnectionCorePlugin.getDefault().getLdifModificationLogger();
+            File[] files = modificationLogger.getFiles( input.getBrowserConnection().getConnection() );
+            modificationLogger.dispose( input.getBrowserConnection().getConnection() );
+            for ( int i = 0; i < files.length; i++ )
             {
-                if ( files[i] != null && files[i].exists() && !files[i].delete() )
+                try
                 {
-                    fw = new FileWriter( files[i] );
-                    fw.write( "" );
+                    if ( files[i] != null && files[i].exists() && !files[i].delete() )
+                    {
+                        fw = new FileWriter( files[i] );
+                        fw.write( "" );
+                    }
+    
                 }
-
+                catch ( Exception e )
+                {
+                    sb.append( e.getMessage() );
+                }
             }
-            catch ( Exception e )
-            {
-                sb.append( e.getMessage() );
-            }
+            view.getMainWidget().getSourceViewer().setTopIndex( 0 );
+            view.getMainWidget().getSourceViewer().getDocument().set( sb.toString() );
         }
-        view.getMainWidget().getSourceViewer().setTopIndex( 0 );
-        view.getMainWidget().getSourceViewer().getDocument().set( sb.toString() );
     }
 
 }
